Broken commit for checkpointing
I was having trouble running this on my NUC, due to illegal opcodes.
Turns out it is due to AVX512 instructions. I didn't realize I was
forcing them to be issued in the top level CMakeLists.txt. So, I need to
do some feature gating on the SIMD instruction front so that I know what
compiles or not.